Correct namespaces after moving file
