+++ src/Infrastructure/Persistence/ApplicationDbContextSeed.cs
@@ -1,6 +1,5 @@
 ﻿using CleanArchitecture.Domain.Entities;
 using CleanArchitecture.Domain.ValueObjects;
-using CleanArchitecture.Infrastructure.Identity;
 using Microsoft.AspNetCore.Identity;
 using System.Linq;
 using System.Threading.Tasks;
