+++ src/WebUI/Pages/Shared/_LoginPartial.cshtml
@@ -1,5 +1,5 @@
 ﻿@using Microsoft.AspNetCore.Identity
-@using CleanArchitecture.Infrastructure.Identity;
+@using CleanArchitecture.Domain.Entities;
 @inject SignInManager<ApplicationUser> SignInManager
 @inject UserManager<ApplicationUser> UserManager
